vfs: add d_canonical_path for stacked filesystem support
Inotify does not currently know when a filesystem
is acting as a wrapper around another fs. This means
that inotify watchers will miss any modifications to
the base file, as well as any made in a separate
stacked fs that points to the same file.
d_canonical_path solves this problem by allowing the fs
to map a dentry to a path in the lower fs. Inotify
can use it to find the appropriate place to watch to
be informed of all changes to a file.

Change-Id: I9b5ad26359460a4adcd600e52d0c2139e82caa3a
Signed-off-by: Daniel Rosenberg <drosen@google.com>
(cherry picked from commit 6c35233197c38905dfa09b9903a27135aafe9abd)